+++ examples/samples/resources/ciscoise_network_access_authorization_rules/resource.tf
@@ -1,7 +1,7 @@
 terraform {
   required_providers {
     ciscoise = {
-      version = "0.6.11-beta"
+      version = "0.6.12-beta"
       source  = "hashicorp.com/edu/ciscoise"
     }
   }
@@ -13,21 +13,48 @@ provider "ciscoise" {
 resource "ciscoise_network_access_authorization_rules" "example" {
   provider = ciscoise
   parameters {
-    profile = ["authp_uc6"]
-    rule {
+    profile = ["Blackhole_Wireless_Access"]
+     rule {
       condition {
-        condition_type = "ConditionReference"
-        name           = "lc_uc6"
-        id             = "716b2dfd-a1e0-4d94-9802-eb4d3bc614d3"
-        is_negate      = "true"
+        condition_type = "ConditionAndBlock"
+        is_negate = false
+        children {
+          condition_type = "ConditionAttributes"
+          is_negate = false
+          dictionary_name = "EndPoints"
+          attribute_name = "LogicalProfile"
+          operator = "equals"
+          attribute_value = "APs"
+        }
+        children {
+          condition_type = "ConditionReference"
+          is_negate = false
+          id = "8bd7ac6e-0910-445e-9d7e-35327cc54e6b"
+        }
       }
-      default = "true"
-      # hit_counts = 1
-      # id         = "716b2dfd-a1e0-4d94-9802-eb4d3bc614d3"
-      name  = "MyRuleName_1"
-      rank  = 1
-      state = "enabled"
-    }
-    security_group = "BYOD"
+     }
+    policy_id="244253d8-cfc9-43b7-8241-38199c71c852"
+    id = "00b8bea1-375e-41fe-acb4-ba5ca368f0e8"
+    # security_group = "BYOD"
   }
 }
+
+# {
+#     "profile": [
+#         "Blackhole_Wireless_Access"
+#     ],
+#     "rule": {
+#         "default": false,
+#         "name": "TestAnsibleIssue81",
+#         "rank": 0,
+#         "state": "enabled",
+#         "condition": {
+#             "conditionType": "ConditionAttributes",
+#             "isNegate": false,
+#             "dictionaryName": "IdentityGroup",
+#             "attributeName": "Name",
+#             "operator": "equals",
+#             "attributeValue": "Endpoint Identity Groups:IAC_Lab1"
+#         }
+#     }
+# }